# Ruby performance improvements