Replaces the accumulators array with a hashmap as sort in Ruby is extremely slow (accounting for 95% of execution time). Also makes a slight improvement to the sort performance

Timings:
| | old | new |
| - | - | - |
| 1 query | 2.3s | 1.25s |
| 50 queries | 63s | 5s |